Add New Packages and Authorization Logic
- Added new dependencies: @next/third-parties, lodash, react-masonry-css, react-virtuoso, uuid.
- Updated DELETE method in route.js to include authorization using @clerk/nextjs/server.
- Changed API pagination limit from 10 to 8 for better user experience.
- Introduced detailed logging in POST method for debugging.
- Added new logo assets to the public directory.

+++ src/app/api/comment/[commentId]/route.js
@@ -2,7 +2,7 @@ import { createErrorResponse } from "@/lib/utils/error";
 import { createResponse } from "@/lib/utils/response";
 import { Comment } from "@/models/comment.model";
 import dbConnect from "@/lib/dbConnect";
-
+import { auth } from "@clerk/nextjs/server";
 // update comment
 export async function PUT(req, { params }) {
   try {
@@ -52,8 +52,12 @@ export async function PUT(req, { params }) {
 // delete comment
 export async function DELETE(req, { params }) {
   try {
+    const user = await auth();
+    const id = user?.sessionClaims?.user_Id;
+
     await dbConnect();
     const { commentId } = await params;
+
     if (!commentId) {
       return createErrorResponse({
         success: false,
@@ -69,6 +73,15 @@ export async function DELETE(req, { params }) {
         message: "Comment not found or Comment ID is invalid",
       });
     }
+
+    if (!comment.commentBy.equals(id)) {
+      return createErrorResponse({
+        success: false,
+        status: 401,
+        message: "You are not authorized to delete this comment",
+      });
+    }
+
     await Comment.findByIdAndDelete(commentId);
     return createResponse({ message: "comment deleted", status: 200 });
   } catch (error) {
